pyproject.toml
[tool.poetry]
name = "tibber-technical-case"
version = "0.1.0"
description = "Tibber Technical case"
authors = ["Nikolai Tatarinov <5min4eq.unity@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
Flask = "^3.0.3"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
